# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: EcoDrishti AI API - Nepal School Climate Intelligence Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /recommendations/generate:
    post:
      operationId: generateRecommendations
      tags: [recommendations]
      summary: Generate new AI recommendations based on latest carbon data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecommendationRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recommendation"
...
  /reports/{id}:
    get:
      operationId: getReport
      tags: [reports]
      summary: Get a specific monthly report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyReport"
  /reports/generate:
    post:
      operationId: generateReport
      tags: [reports]
      summary: Generate a monthly report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportGenerateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyReport"
...
components:
  schemas:
...
    Recommendation:
      type: object
      required: [id, category, title, titleNp, description, descriptionNp, estimatedCarbonReductionKg, difficulty, impact, timeline, status]
      properties:
        id:
          type: integer
        category:
          type: string
          enum: [transport, electricity, water, waste, general]
        title:
          type: string
        titleNp:
          type: string
        description:
          type: string
        descriptionNp:
          type: string
        estimatedCarbonReductionKg:
          type: number
        difficulty:
          type: string
          enum: [easy, medium, hard]
        impact:
          type: string
          enum: [low, medium, high]
        timeline:
          type: string
        status:
          type: string
          enum: [active, completed, dismissed]
        createdAt:
          type: string
          format: date-time
    RecommendationRequest:
      type: object
      required: [submissionId]
      properties:
        submissionId:
          type: integer
    MonthlyReport:
      type: object
      required: [id, month, year, totalEmissionsKg, sustainabilityScore, ecoLeagueRank, highlights, createdAt]
      properties:
        id:
          type: integer
        month:
          type: integer
        year:
          type: integer
        totalEmissionsKg:
          type: number
        sustainabilityScore:
          type: number
        dataConfidenceScore:
          type: number
        ecoLeagueRank:
          type: integer
        highlights:
          type: array
          items:
            type: string
        challengesCompleted:
          type: integer
        activeStudents:
          type: integer
        carbonReductionPercent:
          type: number
        transportEmissionsKg:
          type: number
        electricityEmissionsKg:
          type: number
        waterEmissionsKg:
          type: number
        wasteEmissionsKg:
          type: number
        createdAt:
          type: string
          format: date-time
    ReportGenerateInput:
      type: object
      required: [month, year]
      properties:
        month:
          type: integer
        year:
          type: integer
...